refactor: misc cleanups in example app
